Move converter functions to converting module
