+++ validator/metadata_validation.py
@@ -1,0 +1,3 @@
+
+from converting import data_objects_from_magetab
+
